+++ src/RhinoInside.Revit.GH/Components/Element/HostObject/SlabShape.cs
@@ -157,13 +157,7 @@ namespace RhinoInside.Revit.GH.Components.HostObjects
     {
       if (!Params.GetData(DA, "Host", out Types.HostObject host, x => x.IsValid)) return;
 
-      var shape = default(ARDB.SlabShapeEditor);
-      switch (host)
-      {
-        case Types.Floor floor: shape = floor.Value.SlabShapeEditor; break;
-        case Types.Roof roof:   shape = roof.Value.SlabShapeEditor; break;
-      }
-
+      var shape = host.Value.GetSlabShapeEditor();
       if (shape is null)
       {
         AddRuntimeMessage(GH_RuntimeMessageLevel.Warning, $"Only flat and horizontal slabs, floors or roofs are valid for '{Name}'.");