+++ src/RhinoInside.Revit.GH/Parameters/Input/BuiltInParameters.cs
@@ -23,7 +23,6 @@ namespace RhinoInside.Revit.GH.Parameters.Input
     )
     {
       IconDisplayMode = GH_IconDisplayMode.name;
-      LayoutLevel = 2;
     }
 
     protected override void LoadVolatileData()